requirements.txt
python-telegram-bot == 21.3
python-telegram-bot[job-queue]
requests
beautifulsoup4